move fluid.transpiler.details
